[WEB-3964] refactor: permission layer (#7094)
* refactor: permission layer

* refactor: add original_role to project member serializer

* chore: minor fixes related to permission layer

* fix: strict type checking while checking user permissions

+++ web/core/store/member/base-project-member.store.ts
@@ -1,36 +1,35 @@
 import set from "lodash/set";
 import sortBy from "lodash/sortBy";
 import uniq from "lodash/uniq";
+import unset from "lodash/unset";
 import update from "lodash/update";
 import { action, computed, makeObservable, observable, runInAction } from "mobx";
 import { computedFn } from "mobx-utils";
-// types
-import { EUserPermissions } from "@plane/constants";
-import { IProjectBulkAddFormData, IProjectMember, IProjectMembership, IUserLite } from "@plane/types";
-// plane-web constants
+// plane imports
+import { EUserPermissions, EUserProjectRoles } from "@plane/constants";
+import { IProjectBulkAddFormData, TProjectMembership, IUserLite } from "@plane/types";
+// plane web imports
+import { RootStore } from "@/plane-web/store/root.store";
 // services
 import { ProjectMemberService } from "@/services/project";
 // store
 import { IRouterStore } from "@/store/router.store";
 import { IUserStore } from "@/store/user";
-// store
+// local imports
 import { IProjectStore } from "../project/project.store";
-import { CoreRootStore } from "../root.store";
 import { IMemberRootStore } from ".";
 
-export interface IProjectMemberDetails {
-  id: string;
+export interface IProjectMemberDetails extends Omit<TProjectMembership, "member"> {
   member: IUserLite;
-  role: EUserPermissions;
 }
 
-export interface IProjectMemberStore {
+export interface IBaseProjectMemberStore {
   // observables
   projectMemberFetchStatusMap: {
     [projectId: string]: boolean;
   };
   projectMemberMap: {
-    [projectId: string]: Record<string, IProjectMembership>;
+    [projectId: string]: Record<string, TProjectMembership>;
   };
   // computed
   projectMemberIds: string[] | null;
@@ -39,41 +38,45 @@ export interface IProjectMemberStore {
   getProjectMemberDetails: (userId: string, projectId: string) => IProjectMemberDetails | null;
   getProjectMemberIds: (projectId: string, includeGuestUsers: boolean) => string[] | null;
   // fetch actions
-  fetchProjectMembers: (workspaceSlug: string, projectId: string) => Promise<IProjectMembership[]>;
+  fetchProjectMembers: (
+    workspaceSlug: string,
+    projectId: string,
+    clearExistingMembers?: boolean
+  ) => Promise<TProjectMembership[]>;
   // bulk operation actions
   bulkAddMembersToProject: (
     workspaceSlug: string,
     projectId: string,
     data: IProjectBulkAddFormData
-  ) => Promise<IProjectMembership[]>;
+  ) => Promise<TProjectMembership[]>;
   // crud actions
-  updateMember: (
+  updateMemberRole: (
     workspaceSlug: string,
     projectId: string,
     userId: string,
-    data: { role: EUserPermissions }
-  ) => Promise<IProjectMember>;
+    role: EUserProjectRoles
+  ) => Promise<TProjectMembership>;
   removeMemberFromProject: (workspaceSlug: string, projectId: string, userId: string) => Promise<void>;
 }
 
-export class ProjectMemberStore implements IProjectMemberStore {
+export abstract class BaseProjectMemberStore implements IBaseProjectMemberStore {
   // observables
   projectMemberFetchStatusMap: {
     [projectId: string]: boolean;
   } = {};
   projectMemberMap: {
-    [projectId: string]: Record<string, IProjectMembership>;
+    [projectId: string]: Record<string, TProjectMembership>;
   } = {};
   // stores
   routerStore: IRouterStore;
   userStore: IUserStore;
   memberRoot: IMemberRootStore;
   projectRoot: IProjectStore;
-  rootStore: CoreRootStore;
+  rootStore: RootStore;
   // services
   projectMemberService;
 
-  constructor(_memberRoot: IMemberRootStore, _rootStore: CoreRootStore) {
+  constructor(_memberRoot: IMemberRootStore, _rootStore: RootStore) {
     makeObservable(this, {
       // observables
       projectMemberMap: observable,
@@ -82,7 +85,7 @@ export class ProjectMemberStore implements IProjectMemberStore {
       // actions
       fetchProjectMembers: action,
       bulkAddMembersToProject: action,
-      updateMember: action,
+      updateMemberRole: action,
       removeMemberFromProject: action,
     });
 
@@ -103,6 +106,7 @@ export class ProjectMemberStore implements IProjectMemberStore {
     const projectId = this.routerStore.projectId;
     if (!projectId) return null;
     let members = Object.values(this.projectMemberMap?.[projectId] ?? {});
+    if (members.length === 0) return null;
     members = sortBy(members, [
       (m) => m.member !== this.userStore.data?.id,
       (m) => this.memberRoot.memberMap?.[m.member]?.display_name.toLowerCase(),
@@ -118,19 +122,63 @@ export class ProjectMemberStore implements IProjectMemberStore {
   getProjectMemberFetchStatus = computedFn((projectId: string) => this.projectMemberFetchStatusMap?.[projectId]);
 
   /**
+   * @description get the project memberships
+   * @param projectId
+   */
+  protected getProjectMemberships = computedFn((projectId: string) =>
+    Object.values(this.projectMemberMap?.[projectId] ?? {})
+  );
+
+  /**
+   * @description get the project membership by user id
+   * @param userId
+   * @param projectId
+   */
+  protected getProjectMembershipByUserId = computedFn(
+    (userId: string, projectId: string) => this.projectMemberMap?.[projectId]?.[userId]
+  );
+
+  /**
+   * @description get the role from the project membership
+   * @param userId
+   * @param projectId
+   */
+  protected getRoleFromProjectMembership = computedFn(
+    (userId: string, projectId: string): EUserProjectRoles | undefined => {
+      const projectMembership = this.getProjectMembershipByUserId(userId, projectId);
+      if (!projectMembership) return undefined;
+      const projectMembershipRole = projectMembership.original_role ?? projectMembership.role;
+      return projectMembershipRole ? (projectMembershipRole as EUserProjectRoles) : undefined;
+    }
+  );
+
+  /**
+   * @description Returns the project membership role for a user
+   * @description This method is specifically used when adding new members to a project. For existing members,
+   * the role is fetched directly from the backend during member listing.
+   * @param { string } userId - The ID of the user
+   * @param { string } projectId - The ID of the project
+   * @returns { EUserProjectRoles | undefined } The user's role in the project, or undefined if not found
+   */
+  abstract getUserProjectRole: (userId: string, projectId: string) => EUserProjectRoles | undefined;
+
+  /**
    * @description get the details of a project member
    * @param userId
+   * @param projectId
    */
   getProjectMemberDetails = computedFn((userId: string, projectId: string) => {
-    const projectMember = this.projectMemberMap?.[projectId]?.[userId];
+    const projectMember = this.getProjectMembershipByUserId(userId, projectId);
     if (!projectMember) return null;
     const memberDetails: IProjectMemberDetails = {
       id: projectMember.id,
       role: projectMember.role,
+      original_role: projectMember.original_role,
       member: {
         ...this.memberRoot?.memberMap?.[projectMember.member],
         joining_date: projectMember.created_at,
       },
+      created_at: projectMember.created_at,
     };
     return memberDetails;
   });
@@ -141,7 +189,7 @@ export class ProjectMemberStore implements IProjectMemberStore {
    */
   getProjectMemberIds = computedFn((projectId: string, includeGuestUsers: boolean): string[] | null => {
     if (!this.projectMemberMap?.[projectId]) return null;
-    let members = Object.values(this.projectMemberMap?.[projectId]);
+    let members = this.getProjectMemberships(projectId);
     if (includeGuestUsers === false) {
       members = members.filter((m) => m.role !== EUserPermissions.GUEST);
     }
@@ -158,9 +206,12 @@ export class ProjectMemberStore implements IProjectMemberStore {
    * @param workspaceSlug
    * @param projectId
    */
-  fetchProjectMembers = async (workspaceSlug: string, projectId: string) =>
+  fetchProjectMembers = async (workspaceSlug: string, projectId: string, clearExistingMembers: boolean = false) =>
     await this.projectMemberService.fetchProjectMembers(workspaceSlug, projectId).then((response) => {
       runInAction(() => {
+        if (clearExistingMembers) {
+          unset(this.projectMemberMap, [projectId]);
+        }
         response.forEach((member) => {
           set(this.projectMemberMap, [projectId, member.member], member);
         });
@@ -174,13 +225,17 @@ export class ProjectMemberStore implements IProjectMemberStore {
    * @param workspaceSlug
    * @param projectId
    * @param data
-   * @returns Promise<IProjectMembership[]>
+   * @returns Promise<TProjectMembership[]>
    */
   bulkAddMembersToProject = async (workspaceSlug: string, projectId: string, data: IProjectBulkAddFormData) =>
     await this.projectMemberService.bulkAddMembersToProject(workspaceSlug, projectId, data).then((response) => {
       runInAction(() => {
         response.forEach((member) => {
-          set(this.projectMemberMap, [projectId, member.member], member);
+          set(this.projectMemberMap, [projectId, member.member], {
+            ...member,
+            role: this.getUserProjectRole(member.member, projectId) ?? member.role,
+            original_role: member.role,
+          });
         });
       });
       update(this.projectRoot.projectMap, [projectId, "members"], (memberIds) =>
@@ -195,44 +250,98 @@ export class ProjectMemberStore implements IProjectMemberStore {
 
   /**
    * @description update the role of a member in a project
+   * @param projectId
+   * @param userId
+   * @param role
+   */
+  abstract getProjectMemberRoleForUpdate: (
+    projectId: string,
+    userId: string,
+    role: EUserProjectRoles
+  ) => EUserProjectRoles;
+
+  /**
+   * @description update the role of a member in a project
    * @param workspaceSlug
    * @param projectId
    * @param userId
    * @param data
    */
-  updateMember = async (workspaceSlug: string, projectId: string, userId: string, data: { role: EUserPermissions }) => {
+  updateMemberRole = async (workspaceSlug: string, projectId: string, userId: string, role: EUserProjectRoles) => {
     const memberDetails = this.getProjectMemberDetails(userId, projectId);
-    if (!memberDetails) throw new Error("Member not found");
+    if (!memberDetails || !memberDetails?.id) throw new Error("Member not found");
     // original data to revert back in case of error
-    const originalProjectMemberData = this.projectMemberMap?.[projectId]?.[userId]?.role;
     const isCurrentUser = this.rootStore.user.data?.id === userId;
+    const membershipBeforeUpdate = this.getProjectMembershipByUserId(userId, projectId);
+    const permissionBeforeUpdate = isCurrentUser
+      ? this.rootStore.user.permission.getProjectRoleByWorkspaceSlugAndProjectId(workspaceSlug, projectId)
+      : undefined;
+    const updatedProjectRole = this.getProjectMemberRoleForUpdate(projectId, userId, role);
     try {
       runInAction(() => {
-        set(this.projectMemberMap, [projectId, userId, "role"], data.role);
-        if (isCurrentUser)
-          set(this.rootStore.user.permission.projectUserInfo, [workspaceSlug, projectId, "role"], data.role);
+        set(this.projectMemberMap, [projectId, userId, "original_role"], role);
+        set(this.projectMemberMap, [projectId, userId, "role"], updatedProjectRole);
+        if (isCurrentUser) {
+          set(
+            this.rootStore.user.permission.workspaceProjectsPermissions,
+            [workspaceSlug, projectId],
+            updatedProjectRole
+          );
+        }
+        set(this.rootStore.user.permission.projectUserInfo, [workspaceSlug, projectId, "role"], updatedProjectRole);
       });
       const response = await this.projectMemberService.updateProjectMember(
         workspaceSlug,
         projectId,
         memberDetails?.id,
-        data
+        {
+          role,
+        }
       );
       return response;
     } catch (error) {
       // revert back to original members in case of error
       runInAction(() => {
-        set(this.projectMemberMap, [projectId, userId, "role"], originalProjectMemberData);
-        if (isCurrentUser)
+        set(this.projectMemberMap, [projectId, userId, "original_role"], membershipBeforeUpdate?.original_role);
+        set(this.projectMemberMap, [projectId, userId, "role"], membershipBeforeUpdate?.role);
+        if (isCurrentUser) {
+          set(
+            this.rootStore.user.permission.workspaceProjectsPermissions,
+            [workspaceSlug, projectId],
+            membershipBeforeUpdate?.original_role
+          );
           set(
             this.rootStore.user.permission.projectUserInfo,
             [workspaceSlug, projectId, "role"],
-            originalProjectMemberData
+            permissionBeforeUpdate
           );
+        }
       });
       throw error;
     }
   };
+
+  /**
+   * @description Handles the removal of a member from a project
+   * @param projectId - The ID of the project to remove the member from
+   * @param userId - The ID of the user to remove from the project
+   */
+  protected handleMemberRemoval = (projectId: string, userId: string) => {
+    unset(this.projectMemberMap, [projectId, userId]);
+    set(
+      this.projectRoot.projectMap,
+      [projectId, "members"],
+      this.projectRoot.projectMap?.[projectId]?.members?.filter((memberId) => memberId !== userId)
+    );
+  };
+
+  /**
+   * @description Processes the removal of a member from a project
+   * This abstract method handles the cleanup of member data from the project member map
+   * @param projectId - The ID of the project to remove the member from
+   * @param userId - The ID of the user to remove from the project
+   */
+  abstract processMemberRemoval: (projectId: string, userId: string) => void;
 
   /**
    * @description remove a member from a project
@@ -242,14 +351,11 @@ export class ProjectMemberStore implements IProjectMemberStore {
    */
   removeMemberFromProject = async (workspaceSlug: string, projectId: string, userId: string) => {
     const memberDetails = this.getProjectMemberDetails(userId, projectId);
-    if (!memberDetails) throw new Error("Member not found");
+    if (!memberDetails || !memberDetails?.id) throw new Error("Member not found");
     await this.projectMemberService.deleteProjectMember(workspaceSlug, projectId, memberDetails?.id).then(() => {
       runInAction(() => {
-        delete this.projectMemberMap?.[projectId]?.[userId];
+        this.processMemberRemoval(projectId, userId);
       });
-      this.projectRoot.projectMap[projectId].members = this.projectRoot.projectMap?.[projectId]?.members?.filter(
-        (memberId) => memberId !== userId
-      );
     });
   };
 }
